refactor(e2e): split scenario-evidence.v2 tests into test_evidence_v2.py
CodeScene flagged tests/e2e/scenarios/test_harness.py with a blocking
"Low Cohesion" gate failure (10.00 -> 8.44) after this PR bolted six new
test classes (build identity, evidence-v2 fields, status derivation,
evidence validation, schema-file sync, capability_ids validation — all
about the scenario-evidence.v2 record contract) onto a file whose
existing scope was the harness's step-dispatch/record/sign-off contract.

Moved TestBuildIdentity, TestEvidenceV2Fields, TestStatusDerivation,
TestEvidenceValidation, TestSchemaFileSync, and
TestValidateRunbookCapabilityIds (plus their shared helpers) into a new
test_evidence_v2.py. test_harness.py keeps TestRunScenario, TestRecordRun,
TestRenderSignOff, and TestValidateRunbook — its original, cohesive scope.
No test logic changed; same 97 passed / 5 skipped before and after.
